split files, integrate player control to world
