+++ hello_alex/main.py
@@ -1,0 +1,6 @@
+from display import Display
+
+
+if __name__ == "__main__":
+    d = Display()
+    d.start()
